Repo with Audio
Adding Audio was a lot more difficult than adding an image.  Based on reading Stack Overflow (complicated in itself), there are security requirements with audio.  You can't seem to load audio using JS alone.  I had to adjust HTML.  Then I discovered, I could alter the speed of the audio clip using JS.  It's not supported, so I remapped the clip in an audio editor.  All in all, it was frustrating but educational.